+++ src/test/analysis-format/generate-analysis_test.ts
@@ -92,7 +92,7 @@ suite('generate-analysis', () => {
     suite('generates from package', () => {
       test('does not include external features', async () => {
         const basedir = path.resolve(fixtureDir, 'analysis/bower_packages');
-        const analyzer = Analyzer.createForDirectory(basedir);
+        const analyzer = await Analyzer.createForDirectory(basedir);
         const _package = await analyzer.analyzePackage();
         const metadata = generateAnalysis(_package, analyzer.urlResolver);
         // The fixture only contains external elements
@@ -101,7 +101,7 @@ suite('generate-analysis', () => {
 
       test('includes package features', async () => {
         const basedir = path.resolve(fixtureDir, 'analysis/simple');
-        const analyzer = Analyzer.createForDirectory(basedir);
+        const analyzer = await Analyzer.createForDirectory(basedir);
         const _package = await analyzer.analyzePackage();
         const metadata = generateAnalysis(_package, analyzer.urlResolver);
         assert.equal(metadata.elements && metadata.elements.length, 1);
@@ -185,7 +185,7 @@ function* walkRecursively(dir: string): Iterable<string> {
 }
 
 async function analyzeDir(baseDir: string) {
-  const analyzer = Analyzer.createForDirectory(baseDir);
+  const analyzer = await Analyzer.createForDirectory(baseDir);
   const allFilenames = Array.from(walkRecursively(baseDir));
   const htmlOrJsFilenames =
       allFilenames.filter((f) => f.endsWith('.html') || f.endsWith('.js'));