+++ src/test/javascript/namespace-scanner_test.ts
@@ -21,9 +21,9 @@ import {ScannedNamespace} from '../../javascript/namespace';
 import {NamespaceScanner} from '../../javascript/namespace-scanner';
 import {CodeUnderliner, fixtureDir, runScanner} from '../test-utils';
 
-suite('NamespaceScanner', () => {
+suite('NamespaceScanner', async () => {
   const testFilesDir = path.resolve(fixtureDir, 'namespaces/');
-  const analyzer = Analyzer.createForDirectory(testFilesDir);
+  const analyzer = await Analyzer.createForDirectory(testFilesDir);
   const underliner = new CodeUnderliner(analyzer);
 
   async function getNamespaces(filename: string) {