+++ src/test/polymer/polymer-element_test.ts
@@ -21,9 +21,9 @@ import {Severity, Warning} from '../../model/model';
 import {PolymerElement} from '../../polymer/polymer-element';
 import {fixtureDir} from '../test-utils';
 
-suite('PolymerElement', () => {
+suite('PolymerElement', async () => {
   const analyzer =
-      Analyzer.createForDirectory(path.resolve(fixtureDir, 'polymer2/'));
+      await Analyzer.createForDirectory(path.resolve(fixtureDir, 'polymer2/'));
 
   async function getElements(filename: string): Promise<Set<PolymerElement>> {
     const result = (await analyzer.analyze([filename])).getDocument(filename);